Improve correctness of `incremental` function
Typically we don't want to just roll back the source code that is the
input for the Haskell package because the dependencies for the package
may have changed

In other words, if you roll back the source code for the top-level
package without also rolling back the Nix-supplied dependencies
for that build then you run the risk of an unexpected build failure
(due to an older version of the Haskell package being built against
a newer version of the Nix-supplied dependencies).

What you actually want to do is to roll back the entire repository
(i.e. the Haskell source code and the supporting Nix code) to ensure
that the Haskell source code and Nix code stay in sync.

This more generalized rollback complicates the UX for the
`incremental` function.  I did my best to try to streamline
the UX so that the user just needs to specify how to locate the
matching (older) package after a rollback.