bintools: Add response file support to `ld-wrapper`
The motivation behind this is to alleviate the problem
described in https://github.com/NixOS/nixpkgs/issues/41340.
I'm not sure if this completely fixes the problem, but it
eliminates one more area where we can exceed command line
length limits.

This is essentially the same change as in #112449,
except for `ld-wrapper.sh` instead of `cc-wrapper.sh`.

However, that change alone was not enough; on macOS the
`ld` provided by `darwin.cctools` fails if you use process
substitution to generate the response file, so I put up two
PRs to fix that:

https://github.com/tpoechtrager/cctools-port/pull/131
https://github.com/tpoechtrager/cctools-port/pull/132

… and I included a patch referencing that fix so that the
new `ld-wrapper` still works on macOS.