squawk: init at 0.20.0
Squawk is a useful tool for linting PostgreSQL DDL migrations for
potential downtime risks.

Squawk 0.20.0 was released without updating Cargo.lock, so I patched
that. There was also a test that failed because it snapshots the
internal version of the PostgreSQL parser and there was a minor version
difference. I patched that as well.

Each version of libpg_query corresponds to the PostgreSQL version from
which the parser code is extracted. There are slightly different
abstract syntax trees produced by different versions of PostgreSQL and
Squawk doesn't handle the latest one available in nixpkgs. Override it
just for Squawk.

I pulled this from 45bff9e87be49f8c5b97d5d471039b9242e86251.